requirements.txt
pandas
pyarrow
azure-storage-blob
python-dotenv
numpy